+++ packages/templates/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @react-kino/templates
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - react-kino@0.4.0
+
 ## 0.1.3
 
 ### Patch Changes